+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add experimental native backend support for Windows and Linux ([#1295](https://github.com/getsentry/sentry-unreal/pull/1295))
+- Add stacktrace display to Sentry Crash Reporter window ([#1312](https://github.com/getsentry/sentry-unreal/pull/1312))
 - Add automatic frame time metrics ([#1293](https://github.com/getsentry/sentry-unreal/pull/1293))
 - Add automatic game stats metrics ([#1305](https://github.com/getsentry/sentry-unreal/pull/1305))
 - Add automatic GC pause metrics ([#1302](https://github.com/getsentry/sentry-unreal/pull/1302))